Develop/0.0.6

+++ mall-admin/mall-sys-dao/src/main/java/com/yz/mall/sys/entity/SysUser.java
@@ -63,9 +63,9 @@ public class SysUser extends Model<SysUser> {
     private String password;
 
     /**
-     * 账户余额
+     * 账户余额（单位：分）
      */
-    private BigDecimal balance;
+    private Long balance;
 
     /**
      * 昵称
