+++ mall-pms/mall-pms-dao/src/main/java/com/yz/mall/pms/entity/PmsProduct.java
@@ -9,7 +9,6 @@ import lombok.EqualsAndHashCode;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -63,9 +62,9 @@ public class PmsProduct extends Model<PmsProduct> {
     private String productName;
 
     /**
-     * 商品价格
+     * 商品价格（单位：分）
      */
-    private BigDecimal productPrice;
+    private Long productPrice;
 
     /**
      * 商品标签
