+++ mall-pms/mall-pms-dao/src/main/java/com/yz/mall/pms/vo/PmsProductSlimVo.java
@@ -3,8 +3,6 @@ package com.yz.mall.pms.vo;
 import com.yz.mall.pms.entity.PmsProduct;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 /**
  * 商品信息
  *
@@ -25,9 +23,9 @@ public class PmsProductSlimVo {
     private String productName;
 
     /**
-     * 商品价格
+     * 商品价格（单位：分）
      */
-    private BigDecimal productPrice;
+    private Long productPrice;
 
     /**
      * 商品标签
